refactor: extract BaseKeycap from OverlayKeycapView
Move all base-layer content rendering into a focused BaseKeycap child
view with 23 parameters. Covers: multi-legend (JIS/ISO), novelty keys,
function keys, app launch, URL favicon, system action, inline layer
arrows, centered/bottom-aligned/narrow-modifier/arrow/esc content.

600 lines removed from 4 extension files. Parent retains touchIdContent,
capsLockIndicator, functionKeyWithMappingContent (dots legend), and
computed properties needed by the keycap shell.

Snapshot tests pass pixel-identical. Phase 4.4 of the overlay keycap
decomposition plan.
